requirements.txt
psycopg2-binary
pycaret
streamlit